Add a small layer of logic to not have to rebuild everything when config.status is run.
When config.status is run (which happens when the configure script is modified),
src/inc/maxdim.h and src/inc/maxdimc.h are rewritten. This causes libmir.la to be
rebuilt, which causes everything to be rebuilt, even if the contents of the header files
are unchanged. This patch causes config.status to create src/inc/cfg_maxdim{,c}.h,
which are then copied to maxdim{,c}.h only if necessary. This saves a lot of time if
the configure script changes but the MAXDIM variables don't. More details in comments
in src/inc/Makefile.am.