Import Bob's big-scratchfile and big-invert patches.
We discovered that an INVERT with many visibilities could
fail because the scratch file interface prevented the code
from accessing more than 32 bits' worth of scratchfile,
and an inversion with too many visibilities would exceed
this limit. Bob Sault coded up a workaround in which you
can set a "record size" used by the scratch file routines
and modified INVERT to use this feature. Some preliminary
testing indicates that the feature works and enables
larger inversions.