mirdoc.mak: fix for parallel builds
When running a parallel build, sometimes make would try to build the docs
for a program before the "doc" program had been built. This commit fixes this
by adding in a dependency on the "doc" program for the appropriate places.
We do this by changing from a suffix-style rule to a GNU Make pattern rule.
This is conceivably a portability issue, but I believe we have other
GNU-Make-isms in the build system already that haven't been causing problems.
I'm not sure how else to fix this short of prohibiting parallel builds or
adding a configure check for the capabilities of $MAKE (ugh).