autotools: simplify task-building Makefiles greatly.
We've been relying on GNU make already, so go ahead and take advantage of
its pattern-substitution features to make the task-related Makefile.am's
much shorter. This is enabled by automake's AM_DEFAULT_SOURCE_EXT
variable. One thing is that by using this, we distribute all of the .f
files by default. Work around this by just deleting them in dist-hook
rules; it's not pretty, but it works fine.