autotools: add suport for the silent-rules feature of Automake 1.11
This feature allows the user to make the build output relatively terse
status messages rather than lengthy compiler commandlines. It's enabled
by passing the --enable-silent-rules flag to configure. Hooking it up
is a matter of adding some extra variables to custom rules in Makefile.ams;
for earlier versions of automake, I believe the support is a noop.