src/subs/model.for: use scrrecsz to allow very large scratch files
Update gmakes, gpscal, mselfcal, selfcal, uvmodel, and wsrtuvmodel
to do the same.

This stems from an attempt to selfcal a very large dataset with
a short time interval. The program crashed because the scratchfile
became bigger than 2**32 bytes. This problem was solved by Bob
Sault for invert by adding the scrrecsz API, and that approach
works well in this case too.