autotools: instead of defining @no_f90@, just give ratty -D NOF90
As a side note, the NOF90 conditional is only used once in MIRIAD, in
src/prog/vis/calred.for. So for now it's not worth actually
supporting, even though nowadays we are pretty much always being built
with a F90-compatible compiler. In the future, it'd be nice if that
changed, since dynamic memory allocation is pretty rad.