autotools: make readline optional and only link with it when needed
The main point of this change is to avoid linking most tasks with
readline, since the link can cause binary incompatibilities because
there was a version bump in the readline library.

You need to "make uninstall" before rebuilding for this change to take
effect. Previous libmir.so and friends were built linking with
readline, and libtool will find the installed versions of the
libraries when linking and search their dependencies, causing the
readline dependency to self-perpetuate unless the installed versions
are removed.