autotools: remove -static-libgfortran stuff, didn't work
It turns out that libtool strips out unexpected linker options, so the
-static-libgfortran argument wasn't doing anything.

The libtool-recommended way to pass such flags to the linker is to do
something along the lines of './configure F77="gfortran
-static-libgfortran"', i.e. on a build-by-build basis. This does seem
like the right approach since it works on my Fedora laptop, but not my
SuSE machines.

But the AX_CHECK_COMPILE_FLAG changes are nice, so also deploy them in
pgplot.