autotools: work around complaints from Automake >~ 1.11.3
It doesn't want SCRIPTS or DATA files in the pkglibdir installation
directory. The files I'm installing have platform-specific information
embedded, however, so I think it's a reasonable place to put them.
The documentation says that you should be allowed to put PROGRAMS
in pkglibdir (and LIBRARIES, of course), but this doesn't seem to work
in practice, and if I did this I think it'd try to generate compilation
rules for my scripts. I think this is just silly behavior on the
part of Automake.