borrow/pgplot/cpg/Makefile.am: try to fix parallel builds and distribution
List the outputs of pgbind in BUILT_SOURCES so that they get built
automatically first. This should fix an issue with parallel builds
where sometimes cpgdemo would try to be compiled before cpgplot.h was
built.  This also lets us remove cpgplot.h from the SOURCES of
cpgdemo, which was apparently including cpgplot.h in the distribution,
which it shouldn't be.