mircommon.mak: pass the -a argument to intf2c
This is an attempt to fix the "Some odd form of rounding problem, in mmAlloc"
issue that's been appearing in the macports version of MIRIAD. I can't
reproduce the problem on my very old Mac laptop but it looks like the issue is
64-bit related, and based on my reading of the code (and what Wikipedia says
about the Mac's 64 bit memory model, I32/LP64) we ought to be providing the -a
flag. And it fits in with the MacPorts-specificity of the problem, since
MacPorts uses the autotools build system rather than the classic.