atnf changes from 1.16-1.14:
previous fix introduced new bug, fix that one too

don't check 3rd axis if only 2 requested

In setCel, lonpole and latpole must be reset when changing coordinates,
especially since an inconsistency may result in some cases leading to
a CELERR_ILL_COORD_TRANS error from celset.  In principle they could be
set to give an oblique system so that no regridding need be done.  In
practice, it is assumed that the user wants a map in the "normal"
orientation (north up), and this is clarified in the help.