manuals/carma-cook/carmacookbook.tex: use "rotating" package
I need this to be able to use the "sideways" environment for Figure 6.2. This
fixes compilation for me, but unfortunately the figure is now coming through
upside-down rather than sideways. This is consistent with the documentation in
my version of the rotating package, and it seems that one needs to use a
different environment to be able to control the sense of the rotation (!).