mircommon.mak: remove -DNOF90 define in ratty
I.e., enable F90 constructs. The only place this shows up is in calred.for, a
program that uses enormous arrays that prevent us from being able to build
when various MAX* parameters get large. From the history it's pretty clear
that the F90 dynamic memory allocation was added precisely to work around this
issue.